requirements.txt
flask
gunicorn
python-dotenv
upstash-redis
qrcode[pil]
